requirements.txt
pyserial>=3.4
tqdm>=4.44
